redundancy: remove unspecial special cases (#3755)
As we discussed before, some of the dirs in RAW_DIRNAMES are treated
as standard dirs in an APK, despite not being such.

Instead of arbitrarily deciding what is known and what is unknown,
we can make a concrete distinction: known dirs are dirs that are
standard in APKs as generated by aapt: "assets", "lib" and "res".
Everything else is app or library-specific and is unknown to aapt or
smali/baksmali and should be treated as raw files that must be
preserved as-is.

"libs" (not "lib") is another example of a non-standard dir that might
appear in specific APKs simply because the developers decided so, but
it's not used for native libraries when calling loadLibrary.
https://cs.android.com/android/platform/superproject/main/+/main:frameworks/base/core/java/android/app/LoadedApk.java;l=565

Thanks to refactoring done some months ago, everything in "unknown" is
already being preserved when rebuilding an APK, no special considerations
needed for kotlin.