reorganize some classes and minimize reading signed data from arsc
Move:
* brut.util.AaptManager -> brut.androlib.AaptManager
* brut.androlib.BackgroundWorker -> brut.util.BackgroundWorker
* brut.androlib.meta.Yaml* -> brut.yaml

Removed useless "throws" on signatures of Yaml functions, none of
the Yaml functions throw any managed exceptions, it's a perfectly
generic set of classes that deserves a dedicated module.

Most of data in arsc is encoded as unsigned integers, so normally
we'd use readUnsignedByte/Short where possible (both return int)
to ensure the values are never negative.
Java doesn't have any valid unsigned counterpart for readInt, so
we'll have to live with it, it works for now. Java was never good
for dealing with raw binary data.