refactor: ResChunkPullParser + BinaryDataInputStream
Introducing: ResChunkPullParser

Traverses chunks one by one and provides methods of access to header data along with poisional info.
Inspired by an aapt2 class of the same name:
https://android.googlesource.com/platform/frameworks/base/+/refs/heads/main/tools/aapt2/format/binary/ResChunkPullParser.h

So far we've been parsing ARSC in a linear way, despite ARSC being a nested format (some chunks contain chunks).
This pull parser is the AOSP way of handling this. Each "chunk with chunks" is traversed using it's own pull parser with a size limit to never try to parse past it's data.

In theory, it can be used for BinaryXmlResourceParser too in it's doNext() method, but this method is so messy that
I'd rather study more about these chunks and their quirks before I do any changes here. Maybe in the future.

Introducing: BinaryDataInputStream

Replaces ExtDataInputStream that was basically a matryoshka of an InputStream wrapped in a CountingInputStream wrapped in a DataInputStream/LittleEndianDataInputStream wrapped in a ExtDataInputStream that added extra methods.

BinaryDataInputStream fuses all of the most common features of an InputStream designed to read binary data:

* Our custom methods from ExtDataInputStream.
* Selective byte order (combined DataInputStream and Guava's LittleEndianDataInputStream)
* Tracks position (like Apache/Guava's CountingInputStream).
* Bonus: Maximum read limit (like Apache's BoundedInputStream, or Guava's LimitedInputStream via ByteStreams.limit()).

All in one simple wrapper around the original InputStream.