optimize: track and jump back to last visited location in type chunk
For normal apps, this makes no difference and basically a no-op.
For apps with unordered entries, like Whatsapp, this will avoid false
"Skipped unknown x bytes at end of RES_TABLE_TYPE_TYPE chunk." warnings,
since the stream jumps backwards, and when the entries end the stream is
somewhere in the middle of the data block instead of the end of the last
entry, offset-wise.