Even better - pre-sort by offset
Instead of tracking the last visited position, we just pre-sort
the entry index-to-offset map by offset before we loop it.
The order of the indices doesn't matter, we store the entries
in a HashMap anyway, and sort them before output to XML already.
While index is unique, the offset might repeat for identical
entries , so we mark just before we read an entry, in case we
need to re-read it again next iteration. But overall, the stream
moves forward and by the end of the loop, the stream will be at the
end of the furthermost entry (and the type chunk, unless there are
actual unknown bytes stuffed there).