fix: skip() is skipping if it feels like it
Somebody thought that releasing a method with this comment acceptable:
"The skip method may, for a variety of reasons, end up skipping over some smaller number of bytes, possibly 0."

It's like saying "using this may corrupt your data, have fun".
And this is one reason why Java is crap.

Anyway, this is a rare issue but I encountered it today.
Adding the same workaround we had for skipBytes before.