refactor: ResChunkPullParser + BinaryDataInputStream (#3888)
* refactor: ResChunkPullParser + BinaryDataInputStream

Introducing: ResChunkPullParser

Traverses chunks one by one and provides methods of access to header data along with poisional info.
Inspired by an aapt2 class of the same name:
https://android.googlesource.com/platform/frameworks/base/+/refs/heads/main/tools/aapt2/format/binary/ResChunkPullParser.h

So far we've been parsing ARSC in a linear way, despite ARSC being a nested format (some chunks contain chunks).
This pull parser is the AOSP way of handling this. Each "chunk with chunks" is traversed using it's own pull parser with a size limit to never try to parse past it's data.

In theory, it can be used for BinaryXmlResourceParser too in it's doNext() method, but this method is so messy that
I'd rather study more about these chunks and their quirks before I do any changes here. Maybe in the future.

Introducing: BinaryDataInputStream

Replaces ExtDataInputStream that was basically a matryoshka of an InputStream wrapped in a CountingInputStream wrapped in a DataInputStream/LittleEndianDataInputStream wrapped in a ExtDataInputStream that added extra methods.

BinaryDataInputStream fuses all of the most common features of an InputStream designed to read binary data:

* Our custom methods from ExtDataInputStream.
* Selective byte order (combined DataInputStream and Guava's LittleEndianDataInputStream)
* Tracks position (like Apache/Guava's CountingInputStream).
* Bonus: Maximum read limit (like Apache's BoundedInputStream, or Guava's LimitedInputStream via ByteStreams.limit()).

All in one simple wrapper around the original InputStream.

* tweak log

* correction: ResOverlayables are per ResPackage (not ResTable)

* Catch more specific exception

* Preserve stack traces

* rename: sparseResources to sparseEntries

For consistency with compactEntries and AOSP naming (sparse_entries & compact_entries).

* style

* avoid caching useless IDs

* optimize: track and jump back to last visited location in type chunk

For normal apps, this makes no difference and basically a no-op.
For apps with unordered entries, like Whatsapp, this will avoid false
"Skipped unknown x bytes at end of RES_TABLE_TYPE_TYPE chunk." warnings,
since the stream jumps backwards, and when the entries end the stream is
somewhere in the middle of the data block instead of the end of the last
entry, offset-wise.

* Even better - pre-sort by offset

Instead of tracking the last visited position, we just pre-sort
the entry index-to-offset map by offset before we loop it.
The order of the indices doesn't matter, we store the entries
in a HashMap anyway, and sort them before output to XML already.
While index is unique, the offset might repeat for identical
entries , so we mark just before we read an entry, in case we
need to re-read it again next iteration. But overall, the stream
moves forward and by the end of the loop, the stream will be at the
end of the furthermost entry (and the type chunk, unless there are
actual unknown bytes stuffed there).

* fix: skip() is skipping if it feels like it

Somebody thought that releasing a method with this comment acceptable:
"The skip method may, for a variety of reasons, end up skipping over some smaller number of bytes, possibly 0."

It's like saying "using this may corrupt your data, have fun".
And this is one reason why Java is crap.

Anyway, this is a rare issue but I encountered it today.
Adding the same workaround we had for skipBytes before.