optimize: performance and sanity (#3890)
* optimize: performance and sanity

A quick sanity tweak and some methods renamed to match their aapt2
counterparts for consistency and easier reference.
Some logging tweaks and more permissive chunk parsing to minimize
fatal failures with obfuscated APKs a bit.

Also, parseType has been optimized for performance. Each offset is
processed only once and the parsed value is provided for all of the
entry indexes that pointed to that offset.
Due to this optimization, the stream is now moving purely forward
and never marks/resets, and thus buffering is no longer needed.

* Keep read for non-parsing functions