ensure attrs support formats based on usage
Android doesn't verify that a value encoded in an binary XML was encoded with a format that's supported by the attribute.
The attr's "format" attribute is only enforced by aapt2 during build-time, so some obfuscators manipulate it in a way that doesn't match with actual usage.
Let the binary XML decoder add missing formats whenever necessary.
No change in output for APKs that weren't obfuscated in this specific way.